Added option to set maildigest field and fixed bug
Added --digest option to set the maildigest field in the users table
when createing a user.

Also fixed a problem that seems to have been introduced by an update to
the GetOptionKit library. In the past a blank password could be forced
with --password '', but that now throws a GetOptionKit exception. Added
'magic word' NONE for explicitly setting a blank password (needed for
creating LDAP users)