Fixed course-config-set Command
There was a bug in this command which caused it to always fail ($DB had
not been declared global in the private function to alter the DB).

While fixing the bug I also noticed that the command had debugging
print statements left in place. I replaced those with useful feedback
for users.

Finally, I added a description of the Command to the readme file,
because it was missing.