Make config-set more robust
config-get uses the plugin names 'moodle' and 'core' to refer to the mdl_config table, and all other plugin names to refer to the mdl_config_plugins table. config-set does not behave the same way - it ONLY treates the absense of a plugin as referring to the mdl_config table. This results in inconsistent behaviour between config-get and config-set which can be very confusing.

The addition of a simple if statement removes this confusion.