Create new file CourseRestore.php in Moodle29
The current version off CourseRestore.php resides in Moodle23. It has no feature to overwrite an existing course. Instead the content of the backup is added to the current content of the existing course. This new CourseRestore.php has a new option -o that works together with -e. If used this new option deletes the course content of the existing course before adding the content of the backup file.

This is my first activity in code on GitHub - so please forgive me if I am doing something wrong :-)