questioncategory-create
moosh questioncategory-create --reuse -p 6044 -c 754 -d 清明節 noclass

Creates a question category, 'noclass' in parent category 6044, with
context 754 and info '清明節'. The --reuse option creates the category
if and only if there is not exactly one already in existence with the
same parent, context and name.

Other options are idnumber and infoformat, which defaults to
FORMAT_MARKDOWN.

The command is modeled on (course)category-create and borrows code from
the non-static add_category function of the question_category_object
class, which requires context objects and URLs.

I look forward to extending QuestioncategoryCreate and using
add_category when that class is refactored.