Possibly fixes #324 - fix download-moodle
It appears that the naming convention changed for how Moodle stores
files to be downloaded. I was not involved writing this initial code,
but it's clear that the current code gives a 404 error.

Before, if the version wasn't given as an option, it downloaded the
most recent release. That file was stored like moodle-latest-38.tgz

As of now (3.9), it looks like moodle-latest-39.tgz does not exist,
but moodle-3.9.tgz is indeed the latest version.

Also made some change to this code if a point release wasn't given,
it redirects to moodle-latest-majorminor.tgz that seems to work.
 I don't know how this worked before - I might be missing something.

Also made a couple of Moodle style guideline changes in comments and
'if' spacing.