Simplify/fix resolving of inputfile for config-plugin-import
It was observed, that it was not possible to refer to the config file
to be imported by config-plugin-import with an absolute path.

The existing resolution routine, that converted a relative path to an
absolute path, did not handle this correctly. Instead of rolling our
own routine, this PR replaces it with the built-in function realpath,
which resolves the provided path to a fully qualified one.