add check if uncompresseddir is valid
I removed the deletion of $tempdir content. I needed a clean directory to unpack the plugin to avoid finding the wrong version.php file. I instead create a subdirectory for this purpose.

I thought clearing the $tempdir wouldn't be a problem since the downloaded content was removed after being unpacked anyway.

the check for the validity of uncompresseddir is added